+++ packages/services/hmi-server/src/main/java/software/uncharted/terarium/hmiserver/proxies/simulationservice/SimulationCiemssServiceProxy.java
@@ -7,7 +7,6 @@ import javax.ws.rs.*;
 import javax.ws.rs.core.MediaType;
 import javax.ws.rs.core.Response;
 import software.uncharted.terarium.hmiserver.models.simulationservice.SimulationRequest;
-import software.uncharted.terarium.hmiserver.models.simulationservice.CalibrationRequest;
 import software.uncharted.terarium.hmiserver.models.simulationservice.JobResponse;
